projects/python_screenshot_diff_tracker_openAi_o3_mini/outputs/generated_codebase_UI_no_db_working/requirements.txt
Pillow==9.5.0
opencv-python==4.8.0.76
numpy==1.23.5
